Undo self-foot shooting
The way SABRE_AddTexture previously handled the error
message received from SABRE_PrepareTextureAddition by assigning
to a variable in the test expression part of an if statement
was a bad idea to begin with. The way it was implemented also caused
either error signal value (1, 2) to be changed to 1, effectively
hiding any errors with attempting to grow the texture store.