Add depth ordered rendering
The vertical slices that the view consists of are now rendered
ordered from the farthest one to the closest one. This prepares
the system for the addition of sprites, allowing them to be drawn
as complete sprite images with a single draw_from call as they'll
be drawn at the exact depth where they're in front of everything
they're supposed to be in front of and behind of everything
they should be.

The ordering is made without a separate sorting step, instead
the slices (RenderObjects) are inserted at the correct position
in the doubly linked list right when they're added. This is also
optimized by utilizing the knowledge of the typical input data.
The most common case by far is that each next added slice is
going to be somewhere in the immediate vicinity of the previous
added slice. This is due to the view always consisting of a
series of flat wall surfaces slanted in one way or another and
as the view is built by processing the slices in order from left
to right, each such surface will result in a steadily decreasing
or increasing distance per slice, or rarely a stable, constant
distance over multiple slices.

The optimization then is to simply maintain a pointer to the
last slice added, and start looking for the place for the new
one from its position in the list, instead of from the beginning
of the list.

An arbitrary test frame rendered with this system (at horizontal
resolution 640) resulted in 594 traversal operations between
nodes, whereas without the optimization 72503 operations would
have been needed if each slice was added by starting from the
beginning of the list. So, the optimization cut the number of
operations required down to less than 1%.