Optimize rendering for windows
For each wall (an item in the level array) store a pointer
to the latest added render object associated with that particular
wall. Use that to set the render object list manager's curr
pointer's value before inserting a new render object for that
wall. It is guaranteed that the correct location in the render
object list for that slice is somewhere very near.

When multiple windows are being rendered, this optimization can
reduce the number of list traversal operations by over 98%
(the number is directional and acquired by only a single test,
but due to the nature of the optimization, the percentage will
always be significant when there are windows being rendered).