Implement level edge rendering mode 2
In this mode the raycasting just wraps to the opposite side
of the level and continues from there. A maximum limit of wraps
must be used and is currently set as a variable in the screen
draw actor script. That and other things in the current implementation
require some serious refactoring, the raycast loop is currently an
if else jungle.. :P