Improve level wrap implementation
Also implement level wrapping for projectiles. Currently
a projectile wraps only on collision and only for
projectile-to-wall and projectile-to-floor collisions.

Add a global maximum level wrap count defines as a preprocessor
constant. This allows expanding the level tile to contain a
matching number of render object pointers so that each wrap
depth of a tile has its own pointer to minimize the render object
traversal operations required, greatly boosting the performance.