Add script for generating texture thumbnails
The script uses the MIT licensed scale.bat by Vasil Arnaudov.
The license is included in the scale.bat-LICENSE.txt file.

The script takes all image files in the 'textures' folder and
generates 64x64 sized thumbnails of them and places the generated
thumbnails in the data folder, name in the format '_thumb-0000.png'.