Rewrite gen-txr-thumbs script
The rewritten version relies on a SWE-generated file 'textures.conf'
for locating the texture files that should have thumbnails generated
for them.

To make the system more reliable and easy to use, a set of rules has
been designed concerning the texture file type, file names and the
corresponding animation names in GE. The rules are:

- Texture files must be .png files
- File & animation names must match (but no .png extension in GE animation names)
- File & animation names must be at max 30 characters long
- File & animation names must only include characters A-Z, a-z, 0-9, -, _

The first two rules are not enforceable inside GE, as any script inside GE has
no access to information about the image files associated with specific
animations. The gen-txr-thumbs script will fail if no matching file is found
inside the data folder.

The last two rules are checked for the animation names inside GE, but again
the file names themselves can't be validated.