Fix the crash on first startup after opening the .ged
Also update geui to new version.

The "solution" to the problem was to change the way the styles
are initialized. The old createStyle() function did _something_
weird when setting the title font for the style. The pointer value
seemed to be correct on testing, but dereferencing the pointer
still caused the program to crash from Game Mode back to game editor
main view on the first run after opening the .ged, if one didn't go
and re-load the GEUI page in Global Code.

Setting the font manually also fixed the issue, so there had to be
something going wrong with createStyle() behind the scenes. So far
the best guess is that Game Editor had some kind of trouble either
with aligning (not sure if aligning applied there, though) or with
the amount of values passed by copy instead of by reference (the
whole function returned a Style struct by copy).

The old createStyle() function was replaced with multiple smaller
functions for configuring the style settings. This seems to stop
the crashes. It would still be good to know what exactly caused
the issue, but so far it would seem to be a Game Editor / EiC issue,
and not a problem with the GEUI code per se, even though the code was
also far from good(TM).