Streamline data store usage
Also prevent cases where adding a new item to a data store
could 'secretly' make the shortcut pointer stale. This was due
to realloc having to sometimes allocate a block of memory from
a new address to be able to provide a chunk of memory with the
size requested. The new implementation of the shortcut pointer
requires specifying a function that will be called automatically
on any realloc calls to prevent the pointer from going stale.

Add more textures so that the realloc need is hit, so that the
new implementation can be proven to work.